change type of contraint to unsigned int
git-svn-id: http://svn.code.sf.net/p/rts-2/code/branches/rts-2/REL_0_8_0@5209 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/scheduler/rts2schedule.h
@@ -274,7 +274,7 @@ class Rts2Schedule: public std::vector <Rts2SchedObs*>
 		 *
 		 * @return Number of targets which are infeasible with respect to given constraint.
 		 */
-		double getConstraintFunction (constraintFunc _type)
+		unsigned int getConstraintFunction (constraintFunc _type)
 		{
 			switch (_type)
 			{
